feat: unwrap lsp namespaces to flat exports + barrel
